feat:account CRUD operation

+++ src/entity/Transaction.java
@@ -4,11 +4,13 @@ public class Transaction {
     private int id;
     private String type;
     private Compte compte;
-    private float montant;
+    private double montant;
     private String devise;
     private String categorie;
     private String etiquettes;
     private String date;
     private String hour;
+
+
 
 }
